requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-Bcrypt>=1.0.1
Flask-JWT-Extended>=4.5.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
pandas==2.2.3
psycopg2-binary==2.9.9
Werkzeug==3.0.1
requests==2.31.0
nc-py-api==0.20.2
pillow==11.2.1
pypinindia>=0.1.8
